:recycle: refactor(`streamToBuffer`): Extract `streamToChunks`.
